メッセージ作成イベントを新しいファイルに移動し、記事登録機能をサービスとして分離
